API/requirements.txt
flask
flask-cors
Werkzeug
librosa
numpy
docker
PyJWT
flask-pymongo 